New Artwork Loaders;
Updated the TWEngine by decoupling the loading of assets to now be in separate assemblies from the main game engine.  This will fix the issue of changing the TWEngine's version number or strong name and having to update all the artwork to load into the engine, since now the artwork is loaded using these new "LOADER" assemblies.  This change created 3 new project loaders called 'HeightDataLoader', 'ScenaryDataLoader' and 'InstancedModelLoader'.

Ben Scharbach